Fix ffi security vulnerability
https://github.com/octopusinvitro/octopusinvitro.gitlab.io/network/alert/Gemfile.lock/ffi/open

Remediation

Upgrade ffi to version 1.9.24 or later. For example:

gem "ffi", ">= 1.9.24"

Details:
- CVE-2018-1000201 - moderate severity
- Vulnerable versions: < 1.9.24
- Patched version: 1.9.24

ruby-ffi version 1.9.23 and earlier has a DLL loading issue which can
be hijacked on Windows OS, when a Symbol is used as DLL name instead of
a String This vulnerability appears to have been fixed in v1.9.24 and
later.